delete utils/stellar.ts
